refactor: split hwm
refactor: move common hwm mechanism
